refactor: added popups folder
